% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibm.R
\name{ibm}
\alias{ibm}
\title{A function to estimate an inverse Burr regression model}
\usage{
ibm(outcome, mu = ~1, alpha = ~1, theta = ~1, data = NULL, its = 2000)
}
\arguments{
\item{outcome}{The outcome being modeled.}

\item{mu}{The scale parameter. Default is a constant formula of \code{~ 1} or can be the right-hand side of a formula object with covariates.}

\item{alpha}{The first shape parameter. Default is a constant formula of \code{~ 1} or can be the right-hand side of a formula object with covariates.}

\item{theta}{The second shape parameter. Default is a constant formula of \code{~ 1} or can be the right-hand side of a formula object with covariates.}

\item{data}{The dataset containing the outcome and any covariates. At the moment, the function cannot handle data with missing values.}

\item{its}{The function uses bootstrapping for statistical inference. How many bootstrapped iterations do you want to use? The default is 2000.}
}
\description{
This function allows you to fit an inverse Burr model of some outcome
either using constant parameter values for the distribution's scale and
two shape parameters, or else using covariates to fit these parameters.
}
