% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ibm.R
\name{ll_plot}
\alias{ll_plot}
\title{Plot Pr(X > x) given x for an inverse Burr model}
\usage{
ll_plot(
  mu,
  alpha,
  theta,
  xmin = 0,
  xmax = 1000,
  len = 1000,
  add = F,
  legend = NULL,
  legend_title = NULL
)
}
\arguments{
\item{mu}{The scale parameter for an inverse Burr model.}

\item{alpha}{The first shape parameter for an inverse Burr model.}

\item{theta}{The second shape parameter for an inverse Burr model.}

\item{xmin}{Default is 0. The minimum level of x to show in the plot.}

\item{xmax}{Default is 1000. The maximum level of x to show in the plot.}

\item{len}{Default is 1000. The level of granularity for x.}

\item{add}{Are you adding a new inverse Burr model to an existing plot? Default is \code{FALSE}.}

\item{legend}{Default is \code{NULL}. Do you want to include a legend if adding additional plots? This lets you select the category name for a particular model.}

\item{legend_title}{Default is \code{NULL}. Do you want to provide a custom name for your legend plot?}
}
\description{
This function allows you to plot a
log-log plot of Pr(X > x) given x by specifying values for the three
main parameters for an inverse Burr distribution (scale, shape 1, and shape 2).
}
