% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ibm.R
\name{plot.ibm}
\alias{plot.ibm}
\title{Plot the fit of an inverse Burr model}
\usage{
\method{plot}{ibm}(
  model,
  newdata,
  xmax = 1000,
  len = 1000,
  add = F,
  legend = NULL,
  legend_title = NULL
)
}
\arguments{
\item{model}{An object generated via \code{ibm()}.}

\item{newdata}{A dataset with 1 row of values for model covariates.}

\item{xmax}{Default is 1000. The maximum level of x to show in the plot.}

\item{len}{Default is 1000. The level of granularity for x.}

\item{add}{Are you adding a new inverse Burr model to an existing plot? Default is \code{FALSE}.}

\item{legend}{Default is \code{NULL}. Do you want to include a legend if adding additional plots? This lets you select the category name for a particular model.}

\item{legend_title}{Default is \code{NULL}. Do you want to provide a custom name for your legend plot?}
}
\description{
This function allows you to plot the fit of an inverse Burr model in a
log-log plot. It accepts the output of an \code{ibm()} fitted model and
a required \code{newdata} object to simulate what Pr(X > x) given x conditional
on values for model covariates.
}
